rename files to match case on frontend
