+++ backend/routes/index.js
@@ -1,7 +1,7 @@
 const express = require( 'express' );
 const moment = require( 'moment' );
 const _ = require( 'lodash' );
-const { CalendarClient } = require( '../app/calendar-client' );
+const { CalendarClient } = require( '../app/CalendarClient' );
 
 const assertPresent = ( val ) => {
    if ( typeof val === 'undefined' || val == null || val === '' ) {
